Do refactoring
Split ExpressionTypeAscender into two classes, one for TypeCollector and
one for DataFlowAnalysis
